04_CODE/scripts/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
google-generativeai>=0.5.0
